+++ packages/@kagal-ed25519-secret/src/index.ts
@@ -24,7 +24,9 @@ export {
 } from './signer';
 export {
   asBytes,
+  type Bytes,
   decodeBase64,
   encodeBase64,
+  encodeKey,
   getRandom,
 } from './utils';